+++ client/src/components/table/summary-row/types.d.ts
@@ -3,4 +3,6 @@ import { IRowProps } from 'ka-table/props';
 export type SummaryRowProps = IRowProps & {
   /** Whether the first column should be sticky */
   isFirstColumnSticky: boolean;
+  /** First year for projected data */
+  firstProjectedYear: number;
 };